Minor trolling

+++ app/src/main/java/app/model/media/CD.java
@@ -6,44 +6,48 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class CD extends Media {
-    private Map<String, Long> tracks;
+    private Map<String, Integer> tracks;
 
-    public Map<String, Long> getTracks() {
+    public Map<String, Integer> getTracks() {
         return tracks;
     }
-    public void addTrack(String title, long duration) {
+    public void addTrack(String title, int duration) {
         if (!tracks.containsKey(title)) {
             tracks.put(title, duration);
-        } else {
+        } else
+        {
             //implement logic here
         }
     }
     public void removeTrack(String title) {
         if (tracks.containsKey(title)) {
             tracks.remove(title);
-        } else {
-            //implement logic here
+        } else
+        {
+            //TODO: implement logic here
         }
     }
     public void editTrackTitle(String oldTitle, String newTitle) {
         if (tracks.containsKey(oldTitle)) {
-            long duration = tracks.get(oldTitle);
+            int duration = tracks.get(oldTitle);
             tracks.remove(oldTitle);
             tracks.put(newTitle, duration);
-        } else {
-            //implement logic here
+        } else
+        {
+            //TODO: implement logic here
         }
     }
-    public void editTrackLength(String title, long newDuration) {
+    public void editTrackLength(String title, int newDuration) {
         if (tracks.containsKey(title)) {
             tracks.put(title, newDuration);
-        } else {
-            //implement logic here
+        } else
+        {
+            //TODO: implement logic here
         }
     }
 
     public CD(String title, ArrayList<String> creators, Date releaseDate, ArrayList<String> genres, String language,
-              float price, Map<String, Long> tracks) {
+              float price, Map<String, Integer> tracks) {
         super(title, creators, releaseDate, genres, language, price);
         this.tracks = new HashMap<>(tracks);
     }
